+++ YololEmulator.Tests/Expressions/Num/Or.cs
@@ -12,7 +12,7 @@ namespace YololEmulator.Tests.Expressions.Num
 
             var a = result.GetVariable("a");
 
-            Assert.AreEqual(1, a.Value.Number);
+            Assert.AreEqual(1, (int)a.Value.Number);
         }
 
         [TestMethod]
@@ -22,7 +22,7 @@ namespace YololEmulator.Tests.Expressions.Num
 
             var a = result.GetVariable("a");
 
-            Assert.AreEqual(0, a.Value.Number);
+            Assert.AreEqual(0, (int)a.Value.Number);
         }
 
         [TestMethod]
@@ -33,8 +33,8 @@ namespace YololEmulator.Tests.Expressions.Num
             var a = result.GetVariable("a");
             var b = result.GetVariable("b");
 
-            Assert.AreEqual(1, a.Value.Number);
-            Assert.AreEqual(1, b.Value.Number);
+            Assert.AreEqual(1, (int)a.Value.Number);
+            Assert.AreEqual(1, (int)b.Value.Number);
         }
 
         [TestMethod]
@@ -45,8 +45,8 @@ namespace YololEmulator.Tests.Expressions.Num
             var a = result.GetVariable("a");
             var b = result.GetVariable("b");
 
-            Assert.AreEqual(0, a.Value.Number);
-            Assert.AreEqual(0, b.Value.Number);
+            Assert.AreEqual(0, (int)a.Value.Number);
+            Assert.AreEqual(0, (int)b.Value.Number);
         }
 
         [TestMethod]
@@ -57,8 +57,8 @@ namespace YololEmulator.Tests.Expressions.Num
             var a = result.GetVariable("a");
             var b = result.GetVariable("b");
 
-            Assert.AreEqual(1, a.Value.Number);
-            Assert.AreEqual(1, b.Value.Number);
+            Assert.AreEqual(1, (int)a.Value.Number);
+            Assert.AreEqual(1, (int)b.Value.Number);
         }
 
         [TestMethod]
@@ -69,8 +69,8 @@ namespace YololEmulator.Tests.Expressions.Num
             var a = result.GetVariable("a");
             var b = result.GetVariable("b");
 
-            Assert.AreEqual(0, a.Value.Number);
-            Assert.AreEqual(0, b.Value.Number);
+            Assert.AreEqual(0, (int)a.Value.Number);
+            Assert.AreEqual(0, (int)b.Value.Number);
         }
     }
 }
