+++ YololEmulator.Tests/Expressions/Str/Equality.cs
@@ -12,7 +12,7 @@ namespace YololEmulator.Tests.Expressions.Str
 
             var a = result.GetVariable("a");
 
-            Assert.AreEqual(1, a.Value.Number);
+            Assert.AreEqual(1, (int)a.Value.Number);
         }
 
         [TestMethod]
@@ -24,7 +24,7 @@ namespace YololEmulator.Tests.Expressions.Str
             var b = result.GetVariable("b");
 
             Assert.AreEqual("2", a.Value.String.ToString());
-            Assert.AreEqual(1, b.Value.Number);
+            Assert.AreEqual(1, (int)b.Value.Number);
         }
 
         [TestMethod]
@@ -38,7 +38,7 @@ namespace YololEmulator.Tests.Expressions.Str
 
             Assert.AreEqual("2", a.Value.String.ToString());
             Assert.AreEqual("2", b.Value.String.ToString());
-            Assert.AreEqual(1, c.Value.Number);
+            Assert.AreEqual(1, (int)c.Value.Number);
         }
     }
 }
